Ok, NOW animations are all in ish
kayyyy too bad bout the secondaries though. Might honestly scrap at this time, gotta get proc gen up. But, we'll see how hard it is to attach weps, and if we can't build it for this version, we can build it in the background, and get base architecture to work for it. But I imagine, it'll probably all take intense rigging. I don't even know how rigging works.