JANK level generation
Very Jank level generation so far, buut, it has helped me figure out how I want to generate the levels, and the truth is, I want it to pregenrate, and have the walls of the distaff drop away once selected; we're well on the way to that, and with prefab levels, just a bit of logic and they should be solid. Now, just need enemies, then Items and shtuff afterwards. then it's all details, baby!